requirements.txt
python-dotenv
requests
python-dotenv
streamlit
llama-index-llms-ollama
fastapi
uvicorn